Make personal website a single page.
I decided to remove my "About" page, and place the information that was previously contained there to the index page. I didn't have a lot of content on the site so it just made sense to make this change. Also removed all links in the navbar other than the far right adjusted github/linkedin/email icons.